Moved DatabaseHelper & Added Tournament CRUD
